Arith evaluando expressoes com variaveis
