+++ src/program.txt
@@ -1,9 +1,4 @@
 module program
 
-mut x : float = toFloat ( 1 )
-mut y : int = (2 * 2) ^ (1 + 1)
-mut a : int = abs ( 1 )
-mut h : int = toStr(1)
-
-
-
+mut x : int = 2
+mut y : int = (4^x + 3)
